content/media/test/test_standalone.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Media test: standalone video documents</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body onload="doTest()">

<pre id="test">
<script class="testbody" type="text/javascript">

var iframes = [];

for (var i=0; i<gSmallTests.length; ++i) {
  var test = gSmallTests[i];

  // We can't play WAV files in stand alone documents, so just don't
  // run the test on non-video content types.
  var isVideo = /^video/.test(test.type) ? true : false;
  if (!isVideo || !document.createElement("video").canPlayType(test.type))
    continue;
  
  var f = document.createElement("iframe");
  f.src = test.name;
  f._test = test;
  f.id = "frame" + i;
  iframes.push(f);
  document.body.appendChild(f);
}


function filename(uri) {
  return uri.substr(uri.lastIndexOf("/")+1);
}

function doTest()
{
  for (var i=0; i<iframes.length; ++i) {
    var f = document.getElementById(iframes[i].id);
    var v = f.contentDocument.body.firstChild;
    is(v.tagName.toLowerCase(), "video", "Is video element");
    var src = filename(v.currentSrc);
    is(src, iframes[i]._test.name, "Name ("+src+") should match ("+iframes[i]._test.name+")");
    is(v.controls, true, "Controls set (" + src + ")");
    is(v.autoplay, true, "Autoplay set (" + src + ")");
  }
  SimpleTest.finish();
}

if (iframes.length == 0) {
  todo(false, "No types supported");
} else {
  SimpleTest.waitForExplicitFinish();
}

</script>
</pre>
</body>
</html>
